backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
PyJWT==2.9.0
asyncpg==0.29.0
pydantic==2.8.2
